+++ kuhn/actors.go
@@ -17,7 +17,7 @@ func (chance *Chance) GetId() ActorId {
 type Player struct {
 	Id      ActorId
 	Card    *Card
-	Stack   float64
+	Stack   float32
 	Actions []Action
 }
 
@@ -25,7 +25,7 @@ func (player *Player) GetId() ActorId {
 	return player.Id
 }
 
-func (player *Player) UpdateStack(stack float64) {
+func (player *Player) UpdateStack(stack float32) {
 	player.Stack = stack
 }
 
@@ -45,7 +45,7 @@ func (player *Player) CollectPrivateCard(card *Card) {
 	player.Card = card
 }
 
-func (player *Player) PlaceBet(table *Table, betSize float64) {
+func (player *Player) PlaceBet(table *Table, betSize float32) {
 	table.AddToPot(betSize)
 	player.Stack -= betSize
 }
